refactor(api): extract shared auth middleware to _auth.js
- Add api/_auth.js with requireRole() helper (Vercel ignores _-prefixed files as endpoints)
- Deduplicate ~20 lines of repeated token verification + role check from firebase-write.js and send-groupme.js
- role.js unchanged (fetches name alongside role; different enough to not share)
